Code refactored into a classNeed to code YAML output.
