+++ rviz_simulator/src/synthetic_optimization.cpp
@@ -37,433 +37,17 @@
  */
 
 /* TODO:
- *
+ * 
+ * write inverse transform function
+ * Output to YAML
  * templating the YAML-cpp readers to read automatically/refactoring the save methods
  * Use a YAML conversions class for all other nodes
  * exception handling
  *
  */
 
-#include <ros/ros.h>
-#include <ros/package.h>
+#include "rviz_simulator/camera_calibration_optimizer.h"
 
-#include "ceres/ceres.h"
-#include "ceres/rotation.h"
-
-#include <yaml-cpp/yaml.h>
-
-#include <Eigen/Dense>
-
-// for getting a directory/file listing
-#include <dirent.h>
-
-// for making directories
-#include <sys/stat.h>
-#include <sys/types.h>
-
-#define REFERENCE_T_TARGET_SIZE 6
-#define OBJ_POINTS_SIZE 3
-#define CORNER_POINTS_SIZE 2
-#define NUM_OBJ_POINTS 4
-#define NUM_CORNERS 4
-#define CAMERA_INTRINSICS_SIZE 9
-
-// All transforms are stored as a 6 elemnt double array
-// [0, 1, 2] - rodriguez angle-axis rotation
-// [3, 4, 5] - translation
-
-// The target struct models a fiducial target (E.g. an AprilTag)
-// It stores the world_T_target transform
-// and the 3D object point coordinates in the target frame
-struct Target
-{
-  int targetID;
-  std::array<double, REFERENCE_T_TARGET_SIZE> world_T_target;
-  std::array<std::array<double, OBJ_POINTS_SIZE>, NUM_OBJ_POINTS> obj_points_in_target;
-};
-
-// A detection may occur when a camera takes a picture
-// A dectection consists of a targetID and the pixel [x, y] coordinates of the 4 corners that appear in the image
-// This targetID is looked up in the targets collection to retrieve the corresponding world_t_target
-struct Detection
-{
-  int targetID;
-  std::array<std::array<double, CORNER_POINTS_SIZE>, NUM_CORNERS> corners;
-};
-
-// A picture consists of 0 or more detections
-// The picture struct records the world_T_camera transform which is then inverted to get camera_T_world
-struct Picture
-{
-  std::array<double, REFERENCE_T_TARGET_SIZE> world_T_camera;
-  std::array<double, REFERENCE_T_TARGET_SIZE> camera_T_world;
-  std::vector<Detection> detections;
-};
-
-// Templated functor for the Bundle Adjustment problem
-struct ReProjectionResidual
-{
-  // Constructor
-  ReProjectionResidual(const double* observed_pixel_coordinates, const double* obj_point_in_target)
-  {
-    // projected image pixel coordinates (u,v)
-    this->observed_pixel_coordinates[0] = observed_pixel_coordinates[0];
-    this->observed_pixel_coordinates[1] = observed_pixel_coordinates[1];
-
-    // 3D coordinates in ROSWorld with tag0 as origin (X, Y, Z)
-    this->obj_point_in_target[0] = obj_point_in_target[0];
-    this->obj_point_in_target[1] = obj_point_in_target[1];
-    this->obj_point_in_target[2] = obj_point_in_target[2];
-  }
-
-  // Using the openCV camera calibration model:
-  // https://docs.opencv.org/2.4/modules/calib3d/doc/camera_calibration_and_3d_reconstruction.html
-  template <typename T>
-  bool operator()(const T* const camera_intrinsics, const T* const camera_T_world, const T* const world_T_target,
-                  T* residuals) const
-  {
-    const T point[3] = { T(this->obj_point_in_target[0]), T(this->obj_point_in_target[1]),
-                         T(this->obj_point_in_target[2]) };
-
-    T p[3];
-
-    ceres::AngleAxisRotatePoint(world_T_target, point, p);
-    p[0] += world_T_target[3];
-    p[1] += world_T_target[4];
-    p[2] += world_T_target[5];
-
-    ceres::AngleAxisRotatePoint(camera_T_world, p, p);
-    p[0] += camera_T_world[3];
-    p[1] += camera_T_world[4];
-    p[2] += camera_T_world[5];
-
-    T x_prime = p[0] / p[2];
-    T x_prime_squared = x_prime * x_prime;
-    T y_prime = p[1] / p[2];
-    T y_prime_squared = y_prime * y_prime;
-
-    const T& fx = camera_intrinsics[0];
-    const T& fy = camera_intrinsics[1];
-    const T& cx = camera_intrinsics[2];
-    const T& cy = camera_intrinsics[3];
-    const T& k1 = camera_intrinsics[4];
-    const T& k2 = camera_intrinsics[5];
-    const T& p1 = camera_intrinsics[6];
-    const T& p2 = camera_intrinsics[7];
-    const T& k3 = camera_intrinsics[8];
-    // const T& k4 = camera_intrinsics[9];
-    // const T& k5 = camera_intrinsics[10];
-    // const T& k6 = camera_intrinsics[11];
-    const T k4 = T(0.0);
-    const T k5 = T(0.0);
-    const T k6 = T(0.0);
-
-    T r_raise_2 = x_prime_squared + y_prime_squared;
-    T r_raise_4 = r_raise_2 * r_raise_2;
-    T r_raise_6 = r_raise_4 * r_raise_2;
-
-    T numerator = T(1) + k1 * r_raise_2 + k2 * r_raise_4 + k3 * r_raise_6;
-    T denominator = T(1) + k4 * r_raise_2 + k5 * r_raise_4 + k6 * r_raise_6;
-    T coefficient = numerator / denominator;
-
-    T x_double_prime =
-        x_prime * coefficient + T(2) * p1 * x_prime * y_prime + p2 * (r_raise_2 + T(2) * x_prime_squared);
-    T y_double_prime =
-        y_prime * coefficient + p1 * (r_raise_2 + T(2) * y_prime_squared) + T(2) * p2 * x_prime * y_prime;
-
-    T u = fx * x_double_prime + cx;
-    T v = fy * y_double_prime + cy;
-
-    residuals[0] = u - T(this->observed_pixel_coordinates[0]);
-    residuals[1] = v - T(this->observed_pixel_coordinates[1]);
-
-    return true;
-  }
-
-  // Factory to hide the construction of the CostFunction object from the client code.
-  static ceres::CostFunction* Create(const double* observed_pixel_coordinates, const double* obj_point_in_target)
-  {
-    return (new ceres::AutoDiffCostFunction<ReProjectionResidual, 2, CAMERA_INTRINSICS_SIZE, REFERENCE_T_TARGET_SIZE,
-                                            REFERENCE_T_TARGET_SIZE>(
-        new ReProjectionResidual(observed_pixel_coordinates, obj_point_in_target)));
-  }
-
-private:
-  double observed_pixel_coordinates[2];
-  double obj_point_in_target[3];
-};
-
-// print vector
-template <typename T>
-void printVector(std::vector<T> data)
-{
-  for (T element : data)
-  {
-    ROS_INFO_STREAM(element);
-  }
-  ROS_INFO_STREAM(std::endl);
-}
-
-// print std::array
-template <typename T, std::size_t N>
-void printStdArray(std::array<T, N> data)
-{
-  for (T element : data)
-  {
-    ROS_INFO_STREAM(element);
-  }
-  ROS_INFO_STREAM(std::endl);
-}
-
-void printTargets(std::map<int, Target> targets)
-{
-  std::map<int, Target>::iterator itr;
-  for(itr = targets.begin(); itr != targets.end(); ++itr){
-    //  const YAML::Node& target_node = *target_iterator;
-    const Target target = itr->second;
-    ROS_INFO_STREAM("targetID: " << itr->first);
-    ROS_INFO_STREAM("world_T_target: ");
-    printStdArray(target.world_T_target);
-    // ROS_INFO_STREAM("obj_points_in_target: ");
-    // for (int j = 0; j < target.obj_points_in_target.size(); j++)
-    // {
-    //   ROS_INFO_STREAM("Corner index: " << j);
-    //   printStdArray(target.obj_points_in_target[j]);
-    // }
-  }
-}
-
-// searches a given directory
-// returns the list of detection entites (an entity is a directory or file) with the substring "detections"
-std::vector<std::string> getDetectionEntityNames(std::string directory_path, unsigned char entity_type)
-{
-  std::vector<std::string> detection_entities;
-
-  DIR* dir = opendir(directory_path.c_str());
-  struct dirent* entry = readdir(dir);
-  while (entry != NULL)
-  {
-    if (entry->d_type == entity_type)
-    {
-      std::string name = entry->d_name;
-      if (name.find("detections") != std::string::npos)  // a directory
-      {
-        detection_entities.push_back(name);
-      }
-    }
-    entry = readdir(dir);
-  }
-  return detection_entities;
-}
-
-std::array<double, 6> getReference_T_Target(const YAML::Node& node, std::string reference_T_target_name, bool inverse)
-{
-  std::array<double, 6> reference_T_target;
-
-  if (inverse)
-  {
-    // getting rodrigues angle axis rotation and converting to Eigen::Matrix3d
-    std::array<double, 3> rodrigues_angle_axis = node[reference_T_target_name]["rotation"].as<std::array<double, 3>>();
-    double rotation_array[9];
-    ceres::AngleAxisToRotationMatrix(rodrigues_angle_axis.data(), rotation_array);
-    Eigen::Matrix3d rotation_matrix = Eigen::Map<Eigen::Matrix3d>(rotation_array);
-
-    // getting translation and converting to Eigen::Vector3d
-    std::array<double, 3> translation = node[reference_T_target_name]["translation"].as<std::array<double, 3>>();
-    Eigen::Vector3d translation_vector = Eigen::Map<Eigen::Vector3d>(translation.data());
-
-    // combining rotation matrix and translation vector to Eigen::Affine3d
-    Eigen::Affine3d transform;
-    transform.translation() = translation_vector;
-    transform.linear() = rotation_matrix;
-
-    // inverting the transform
-    Eigen::Affine3d transform_inverse = transform.inverse();
-
-    // converting from rotation matrix back to rodriguez angle axis
-    ceres::RotationMatrixToAngleAxis(transform_inverse.rotation().data(), rodrigues_angle_axis.data());
-
-    // storing in an std::array
-    reference_T_target[0] = rodrigues_angle_axis[0];
-    reference_T_target[1] = rodrigues_angle_axis[1];
-    reference_T_target[2] = rodrigues_angle_axis[2];
-
-    reference_T_target[3] = transform_inverse.translation().x();
-    reference_T_target[4] = transform_inverse.translation().y();
-    reference_T_target[5] = transform_inverse.translation().z();
-  }
-  else
-  {
-    reference_T_target[0] = node[reference_T_target_name]["rotation"][0].as<double>();
-    reference_T_target[1] = node[reference_T_target_name]["rotation"][1].as<double>();
-    reference_T_target[2] = node[reference_T_target_name]["rotation"][2].as<double>();
-
-    reference_T_target[3] = node[reference_T_target_name]["translation"][0].as<double>();
-    reference_T_target[4] = node[reference_T_target_name]["translation"][1].as<double>();
-    reference_T_target[5] = node[reference_T_target_name]["translation"][2].as<double>();
-  }
-
-  return reference_T_target;
-}
-
-// returns a single detection from the YAML file
-Detection getDetection(const YAML::Node& detection_node)
-{
-  Detection detection;
-  detection.targetID = detection_node["targetID"].as<int>();
-  for (YAML::const_iterator corner_iterator = detection_node["corners"].begin();
-       corner_iterator != detection_node["corners"].end(); ++corner_iterator)
-  {
-    int corner_index = corner_iterator->first.as<int>();
-    detection.corners[corner_index] = corner_iterator->second.as<std::array<double, CORNER_POINTS_SIZE>>();
-  }
-  return detection;
-}
-
-// returns a vector of all detections from a given YAML file
-std::vector<Detection> getDetections(const YAML::Node& detections_node)
-{
-  std::vector<Detection> detections;
-  for (YAML::const_iterator detection_iterator = detections_node.begin(); detection_iterator != detections_node.end();
-       ++detection_iterator)
-  {
-    detections.push_back(getDetection(*detection_iterator));
-  }
-  return detections;
-}
-
-// returns a picture from a single YAML file
-Picture getPicture(std::string picture_file_path)
-{
-  YAML::Node picture_node = YAML::LoadFile(picture_file_path);
-  if (!picture_node)
-  {
-    ROS_ERROR("Picture file read error.");
-  }
-
-  Picture picture;
-  picture.world_T_camera = getReference_T_Target(picture_node, "world_T_camera", false);
-  picture.camera_T_world = getReference_T_Target(picture_node, "world_T_camera", true);
-  picture.detections = getDetections(picture_node["detections"]);
-  return picture;
-}
-
-// returns a vector of all pictures from a given directory
-// if there are no detections in a picture, then that picture is not pushed to the vector
-std::vector<Picture> getPictures(std::string detections_directory_path)
-{
-  // get list of detection files
-  std::vector<std::string> detection_file_names = getDetectionEntityNames(detections_directory_path, DT_REG);
-
-  std::vector<Picture> pictures;
-  for (auto detection_file_name : detection_file_names)
-  {
-    Picture picture = getPicture(detections_directory_path + "/" + detection_file_name);
-    if (!picture.detections.empty())
-    {
-      pictures.push_back(picture);
-      ROS_INFO_STREAM("Picture added: " << detection_file_name);
-    }
-    else
-    {
-      ROS_INFO_STREAM("No detections in picture: " << detection_file_name);
-    }
-  }
-  return pictures;
-}
-
-void printPictures(std::vector<Picture> pictures)
-{
-  for (int i = 0; i < pictures.size(); i++)
-  {
-    ROS_INFO_STREAM("Picture #" << i);
-    ROS_INFO("world_T_camera");
-    printStdArray(pictures[i].world_T_camera);
-    ROS_INFO("camera_T_world");
-    printStdArray(pictures[i].camera_T_world);
-    for (Detection detection : pictures[i].detections)
-    {
-      ROS_INFO_STREAM("targetID: " << detection.targetID);
-      for (int j = 0; j < detection.corners.size(); j++)
-      {
-        ROS_INFO_STREAM("Corner #" << j << ": ");
-        printStdArray(detection.corners[j]);
-      }
-    }
-  }
-  ROS_INFO("\n");
-}
-
-// reads target data from a YAML file
-// returns a vector of targets
-std::map<int, Target> getTargets(std::string targets_directory_path)
-{
-  YAML::Node targets_node = YAML::LoadFile(targets_directory_path);
-  if (!targets_node)
-  {
-    ROS_ERROR("Targets file read error.");
-  }
-
-  std::map<int, Target> targets;
-
-  // iterating through all targets
-  for (YAML::const_iterator target_iterator = targets_node["targets"].begin();
-       target_iterator != targets_node["targets"].end(); ++target_iterator)
-  {
-    const YAML::Node& target_node = *target_iterator;
-    Target target;
-
-    // targetID
-    target.targetID = target_node["targetID"].as<int>();
-
-    // world_T_target
-    target.world_T_target = getReference_T_Target(target_node, "world_T_target", false);
-
-    // obj_points_in_target
-    for (YAML::const_iterator obj_points_iterator = target_node["obj_points_in_target"].begin();
-         obj_points_iterator != target_node["obj_points_in_target"].end(); ++obj_points_iterator)
-    {
-      int corner_index = obj_points_iterator->first.as<int>();
-      target.obj_points_in_target[corner_index] = obj_points_iterator->second.as<std::array<double, OBJ_POINTS_SIZE>>();
-    }
-
-    targets.insert(std::pair<int, Target>(target.targetID, target));
-  }
-
-  return targets;
-}
-
-// returns an array of camera intrinsics
-std::array<double, CAMERA_INTRINSICS_SIZE> getCameraIntrinsics(std::string camera_file_path)
-{
-  YAML::Node camera_node = YAML::LoadFile(camera_file_path);
-  if (!camera_node)
-  {
-    ROS_ERROR("Camera intrisics file read error.");
-  }
-
-  std::array<double, CAMERA_INTRINSICS_SIZE> camera_intrinsics;
-
-  // XXX: hardcoded for the plumb_bob distortion model
-  camera_intrinsics[0] = camera_node["camera_matrix"]["data"][0].as<double>();  // fx
-  camera_intrinsics[1] = camera_node["camera_matrix"]["data"][4].as<double>();  // fy
-  camera_intrinsics[2] = camera_node["camera_matrix"]["data"][2].as<double>();  // cx
-  camera_intrinsics[3] = camera_node["camera_matrix"]["data"][5].as<double>();  // cy
-
-  // plumb_bob distortions
-  camera_intrinsics[4] = camera_node["distortion_coefficients"]["data"][0].as<double>();  // k1
-  camera_intrinsics[5] = camera_node["distortion_coefficients"]["data"][1].as<double>();  // k2
-  camera_intrinsics[6] = camera_node["distortion_coefficients"]["data"][2].as<double>();  // p1
-  camera_intrinsics[7] = camera_node["distortion_coefficients"]["data"][3].as<double>();  // p2
-  camera_intrinsics[8] = camera_node["distortion_coefficients"]["data"][4].as<double>();  // k3
-  // camera_intrinsics[9] = 0;                                                               // k4
-  // camera_intrinsics[10] = 0;                                                              // k5
-  // camera_intrinsics[11] = 0;                                                              // k6
-
-  return camera_intrinsics;
-}
-
-// gets the directory path that contains all the picture files (E.g. detection_0.yaml)
 std::string getDetectionsDirectoryPath(ros::NodeHandle& n)
 {
   // checking if the detections directory name was passed as a cmd line arg
@@ -487,79 +71,15 @@ std::string getDetectionsDirectoryPath(ros::NodeHandle& n)
   return detections_directory_path;
 }
 
-// optimized results consists of 
-// - camera_intrinsics
-// - world_T_camera for each image
-// - world_T_target for each fiducial target
-
-// function prints optimized results to console
-void printOptimizedResults()
-{
-
-}
-
-// function outputs optimized results to YAML files created in a new directory 
-void optimizedResultsToYAML(std::string output_directory_path)
-{
-  YAML::Node camera_node = YAML::LoadFile(output_directory_path);
-  if (!camera_node)
-  {
-    ROS_ERROR("Camera intrisics file read error.");
-  }
-
-}
-
 int main(int argc, char** argv)
 {
   ros::init(argc, argv, "synthetic_optimization");
   ros::NodeHandle n("~");
 
-  std::string detections_directory_path = getDetectionsDirectoryPath(n);
-
-  std::array<double, CAMERA_INTRINSICS_SIZE> camera_intrinsics =
-      getCameraIntrinsics(detections_directory_path + "/camera.yaml");
-  std::array<double, CAMERA_INTRINSICS_SIZE> initial_intrinsics = camera_intrinsics;
-
-  std::map<int, Target> targets = getTargets(detections_directory_path + "/targets.yaml");
-  // printTargets(targets);
-
-
-  std::vector<Picture> pictures = getPictures(detections_directory_path);
-
-  ceres::Problem problem;
-
-  for (int p = 0; p < pictures.size(); p++)
-  {
-    for (int d = 0; d < pictures[p].detections.size(); d++)
-    {
-      for (int c = 0; c < pictures[p].detections[d].corners.size(); c++)
-      {
-        ceres::CostFunction* cost_function =
-            ReProjectionResidual::Create(pictures[p].detections[d].corners[c].data(),
-                                         targets.at(pictures[p].detections[d].targetID).obj_points_in_target[c].data());
-
-        problem.AddResidualBlock(cost_function, NULL, camera_intrinsics.data(), pictures[p].camera_T_world.data(),
-                                 targets.at(pictures[p].detections[d].targetID).world_T_target.data());
-      }
-    }
-  }
-
-  ceres::Solver::Options options;
-  options.linear_solver_type = ceres::DENSE_SCHUR;
-  options.minimizer_progress_to_stdout = true;
-  options.max_num_iterations = 10000;
-  ceres::Solver::Summary summary;
-  ceres::Solve(options, &problem, &summary);
-  ROS_INFO_STREAM(summary.FullReport() << "\n");
-  std::cout << "\n";
-
-  ROS_INFO_STREAM("INITIAL\t\tFINAL");
-  for (int i = 0; i < camera_intrinsics.size(); i++)
-  {
-    ROS_INFO_STREAM(initial_intrinsics[i] << "\t\t" << camera_intrinsics[i]);
-  }
-
-  printTargets(targets);
-
+  std::string detections_directory_name = getDetectionsDirectoryPath(n);
+  camera_calibration::CameraCalibrationOptimizer camera_calibration_optimizer(detections_directory_name);
+  camera_calibration_optimizer.optimize();
+  camera_calibration_optimizer.printResultsToConsole();
+  
   return 0;
 }
